fix: Ajustado movimenta~ao

+++ Sprint-2/Sprint 2/Assets/Scripts/Object/CharacterModules/Attribute.cs
@@ -7,22 +7,22 @@ using UnityEngine;
 [Serializable]
 public class Attribute
 {
-	public int Current;
-	public int Start;
+	public float Current;
+	public float Start;
 	public float Percentage => (float)Current / (float)Start;
 
-	public Attribute(int start)
+	public Attribute(float start)
 	{
 		Current = Start = start;
 	}
 
-	public bool Remove(int amount)
+	public bool Remove(float amount)
 	{
 		Current -= amount;
 		return Current <= 0;
 	}
 
-	public bool Add(int amount)
+	public bool Add(float amount)
 	{
 		Current += amount;
 		Current = Math.Min(Current, Start);
